+++ src/jwodder_ps1/__init__.py
@@ -20,7 +20,7 @@ Features:
 Visit <https://github.com/jwodder/ps1.py> for more information.
 """
 
-__version__ = "0.6.0.dev1"
+__version__ = "0.6.0"
 __author__ = "John Thorvald Wodder II"
 __author_email__ = "ps1@varonathe.org"
 __license__ = "MIT"